+++ sentry/src/test/java/io/sentry/SentryThreadFactoryTest.kt
@@ -41,13 +41,19 @@ class SentryThreadFactoryTest {
     @Test
     fun `when currentThreads is called, some thread stack frames are captured`() {
         val sut = fixture.getSut()
-        assertTrue(sut.getCurrentThreads(null)!!.filter { it.stacktrace != null }.any { it.stacktrace.frames.count() > 0 })
+        assertTrue(sut.getCurrentThreads(null)!!.filter { it.stacktrace != null }.any { it.stacktrace.frames!!.count() > 0 })
+    }
+
+    @Test
+    fun `when currentThreads is called, stack traces are snapshot`() {
+        val sut = fixture.getSut()
+        assertTrue(sut.getCurrentThreads(null)!!.filter { it.stacktrace != null }.any { it.stacktrace.snapshot == true })
     }
 
     @Test
     fun `when currentThreads and attachStacktrace is disabled, stack frames are not captured`() {
         val sut = fixture.getSut(false)
-        assertFalse(sut.getCurrentThreads(null)!!.filter { it.stacktrace != null }.any { it.stacktrace.frames.count() > 0 })
+        assertFalse(sut.getCurrentThreads(null)!!.filter { it.stacktrace != null }.any { it.stacktrace.frames!!.count() > 0 })
     }
 
     @Test